Refactor sampling related code.
